# api/openapi.yml
openapi: 3.0.3
info:
  title: Developer Utilities MCP Server
  description: |
    A comprehensive, high-performance utility server that consolidates essential development tools into a single service. 
    This server combines cryptography, data manipulation, text processing, identifier generation, time utilities, 
    and network tools to provide AI agents and developers with a one-stop solution for common development tasks.
    
    The server is designed to be secure, stateless, and highly performant, following enterprise-grade standards 
    while remaining accessible for open-source development.
  version: 1.0.0
  contact:
    name: Developer Utilities MCP Server
    url: https://github.com/example/dev-utilities
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/v1/crypto/hash:
    post:
      tags:
        - Cryptography
      summary: Calculate hash
      description: Calculate hash using various algorithms (MD5, SHA1, SHA256, SHA512)
      operationId: calculateHash
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HashRequest'
            examples:
              sha256:
                summary: SHA256 hash example
                value:
                  content: "hello world"
                  algorithm: "sha256"
              md5:
                summary: MD5 hash example
                value:
                  content: "test data"
                  algorithm: "md5"
      responses:
        '200':
          description: Hash calculated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HashResponse'
              examples:
                sha256:
                  summary: SHA256 hash result
                  value:
                    success: true
                    data:
                      hash: "b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9"
                      algorithm: "sha256"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/crypto/hmac:
    post:
      tags:
        - Cryptography
      summary: Generate HMAC
      description: Generate Hash-based Message Authentication Code using SHA256 or SHA512
      operationId: generateHMAC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HMACRequest'
            examples:
              sha256:
                summary: HMAC-SHA256 example
                value:
                  content: "what do ya want for nothing?"
                  key: "Jefe"
                  algorithm: "sha256"
      responses:
        '200':
          description: HMAC generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HMACResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/crypto/password/hash:
    post:
      tags:
        - Cryptography
      summary: Hash password
      description: Hash password using Argon2id with secure defaults
      operationId: hashPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordHashRequest'
            examples:
              basic:
                summary: Basic password hashing
                value:
                  password: "mySecurePassword123"
      responses:
        '200':
          description: Password hashed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PasswordHashResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/crypto/password/verify:
    post:
      tags:
        - Cryptography
      summary: Verify password
      description: Verify password against Argon2id hash with constant-time comparison
      operationId: verifyPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordVerifyRequest'
            examples:
              valid:
                summary: Valid password verification
                value:
                  password: "mySecurePassword123"
                  hash: "$argon2id$v=19$m=65536,t=3,p=4$..."
      responses:
        '200':
          description: Password verification completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PasswordVerifyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/v1/crypto/cert/decode:
    post:
      tags:
        - Cryptography
      summary: Decode X.509 certificate
      description: Decode PEM-encoded X.509 certificate and extract information
      operationId: decodeCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateDecodeRequest'
            examples:
              basic:
                summary: Certificate decoding
                value:
                  certificate: "-----BEGIN CERTIFICATE-----\nMIID...\n-----END CERTIFICATE-----"
      responses:
        '200':
          description: Certificate decoded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CertificateDecodeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Text Processing Endpoints
...
  /api/v1/id/uuid:
    post:
      tags:
        - Identifiers
      summary: Generate UUIDs
      description: Generate UUID version 1 or 4 with configurable count
      operationId: generateUUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UUIDRequest'
            examples:
              uuid_v4:
                summary: Generate UUID v4
                value:
                  version: 4
                  count: 5
              uuid_v1:
                summary: Generate UUID v1
                value:
                  version: 1
                  count: 1
      responses:
        '200':
          description: UUIDs generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UUIDResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/id/nanoid:
    post:
      tags:
        - Identifiers
      summary: Generate Nano IDs
      description: Generate URL-friendly Nano IDs with configurable size and count
      operationId: generateNanoID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NanoIDRequest'
            examples:
              default:
                summary: Default Nano ID
                value:
                  size: 21
                  count: 5
              custom:
                summary: Custom size Nano ID
                value:
                  size: 10
                  count: 3
      responses:
        '200':
          description: Nano IDs generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/NanoIDResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

  # Time Utilities Endpoints
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_ERROR
            - AUTHORIZATION_ERROR
            - RATE_LIMIT_ERROR
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request format"
        details:
          type: string
          description: Additional error details
          example: "Field 'algorithm' is required"

    # Cryptography Schemas
    HashRequest:
      type: object
      required:
        - content
        - algorithm
      properties:
        content:
          type: string
          description: Content to hash
          example: "hello world"
        algorithm:
          type: string
          description: Hash algorithm to use
          enum: [md5, sha1, sha256, sha512]
          example: "sha256"
    HashResponse:
      type: object
      required:
        - hash
        - algorithm
      properties:
        hash:
          type: string
          description: Calculated hash in hexadecimal format
          example: "b94d27b9934d3e08a52e52d7da7dabfac484efe37a5380ee9088f7ace2efcde9"
        algorithm:
          type: string
          description: Algorithm used for hashing
          example: "sha256"
    HMACRequest:
      type: object
      required:
        - content
        - key
        - algorithm
      properties:
        content:
          type: string
          description: Content to authenticate
          example: "what do ya want for nothing?"
        key:
          type: string
          description: Secret key for HMAC
          example: "Jefe"
        algorithm:
          type: string
          description: HMAC algorithm to use
          enum: [sha256, sha512]
          example: "sha256"
    HMACResponse:
      type: object
      required:
        - hmac
        - algorithm
      properties:
        hmac:
          type: string
          description: Calculated HMAC in hexadecimal format
          example: "5bdcc146bf60754e6a042426089575c75a003f089d2739839dec58b964ec3843"
        algorithm:
          type: string
          description: Algorithm used for HMAC
          example: "sha256"
    PasswordHashRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: Password to hash
          minLength: 1
          example: "mySecurePassword123"
    PasswordHashResponse:
      type: object
      required:
        - hash
      properties:
        hash:
          type: string
          description: Argon2id hash of the password
          example: "$argon2id$v=19$m=65536,t=3,p=4$..."
    PasswordVerifyRequest:
      type: object
      required:
        - password
        - hash
      properties:
        password:
          type: string
          description: Password to verify
          example: "mySecurePassword123"
        hash:
          type: string
          description: Argon2id hash to verify against
          example: "$argon2id$v=19$m=65536,t=3,p=4$..."
    PasswordVerifyResponse:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
          description: Whether the password matches the hash
          example: true
    CertificateDecodeRequest:
      type: object
      required:
        - certificate
      properties:
        certificate:
          type: string
          description: PEM-encoded X.509 certificate
          example: "-----BEGIN CERTIFICATE-----\nMIID...\n-----END CERTIFICATE-----"
    CertificateDecodeResponse:
      type: object
      required:
        - certificate
      properties:
        certificate:
          $ref: '#/components/schemas/CertificateInfo'
    CertificateInfo:
      type: object
      required:
        - subject
        - issuer
        - notBefore
        - notAfter
        - serialNumber
        - version
        - keyUsage
        - dnsNames
      properties:
        subject:
          type: string
          description: Certificate subject
          example: "CN=example.com,O=Example Corp,C=US"
        issuer:
          type: string
          description: Certificate issuer
          example: "CN=Example CA,O=Example Corp,C=US"
        notBefore:
          type: string
          format: date-time
          description: Certificate validity start date
          example: "2023-01-01T00:00:00Z"
        notAfter:
          type: string
          format: date-time
          description: Certificate validity end date
          example: "2024-01-01T00:00:00Z"
        serialNumber:
          type: string
          description: Certificate serial number
          example: "1234567890"
        version:
          type: integer
          description: Certificate version
          example: 3
        keyUsage:
          type: array
          items:
            type: string
          description: Certificate key usage
          example: ["Digital Signature", "Key Encipherment"]
        dnsNames:
          type: array
          items:
            type: string
          description: Subject Alternative Names (DNS)
          example: ["example.com", "www.example.com"]

    # Text Processing Schemas
...
    UUIDRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: integer
          description: UUID version
          enum: [1, 4]
          example: 4
        count:
          type: integer
          description: Number of UUIDs to generate
          minimum: 1
          maximum: 1000
          default: 1
          example: 5
    UUIDResponse:
      type: object
      required:
        - uuids
      properties:
        uuids:
          type: array
          items:
            type: string
            format: uuid
          description: Generated UUIDs
          example: 
            - "123e4567-e89b-12d3-a456-426614174000"
            - "987fcdeb-51a2-43d1-9f12-345678901234"
    NanoIDRequest:
      type: object
      properties:
        size:
          type: integer
          description: Size of Nano ID
          minimum: 1
          maximum: 50
          default: 21
          example: 21
        count:
          type: integer
          description: Number of Nano IDs to generate
          minimum: 1
          maximum: 1000
          default: 1
          example: 5
    NanoIDResponse:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: string
          description: Generated Nano IDs
          example: 
            - "V1StGXR8_Z5jdHi6B-myT"
            - "3PIlw-fEXFE1-9Aj-3NNj"

    # Time Utilities Schemas
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_field:
              summary: Missing required field
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' is required"
            invalid_value:
              summary: Invalid field value
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' must be one of: md5, sha1, sha256, sha512"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            generic:
              summary: Generic internal error
              value:
                success: false
                error:
                  code: "INTERNAL_ERROR"
                  message: "Internal server error"
                  details: "An unexpected error occurred"
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication